+++ NEWS.md
@@ -1,3 +1,10 @@
+# flextable 0.6.4
+
+## new features
+
+* `set_flextable_defaults` now accept argument `padding` that set values for padding 
+top, bottom, left and right.
+
 # flextable 0.6.3
 
 ## new features